backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
docxtpl
google-generativeai
python-dotenv
PyMuPDF
Pillow